+++ frontend/src/components/Index/GroupServiceFailures.vue
@@ -35,6 +35,9 @@ export default {
   computed: {
     service_txt() {
       if (!this.service.online) {
+        if (!this.toUnix(this.service.last_success)) {
+          return `Always Offline`
+        }
         return `Offline for ${this.ago(this.service.last_success)}`
       }
       return `${this.service.online_24_hours}% Uptime`