+++ types/core/database.go
@@ -47,7 +47,6 @@ func (c *Core) Create() error {
 		Name:        c.Name,
 		Description: c.Description,
 		ConfigFile:  utils.Directory + "/config.yml",
-		ApiKey:      utils.RandomString(32),
 		ApiSecret:   secret,
 		Version:     App.Version,
 		Domain:      c.Domain,