+++ types/null/unmarshal.go
@@ -29,3 +29,43 @@ func (s *NullString) UnmarshalJSON(b []byte) error {
 	s.Valid = (err == nil)
 	return err
 }
+
+// UnmarshalYAML for NullInt64
+func (i *NullInt64) UnmarshalYAML(unmarshal func(interface{}) error) error {
+	var val int64
+	if err := unmarshal(&val); err != nil {
+		return err
+	}
+	*i = NewNullInt64(val)
+	return nil
+}
+
+// UnmarshalYAML for NullFloat64
+func (f *NullFloat64) UnmarshalYAML(unmarshal func(interface{}) error) error {
+	var val float64
+	if err := unmarshal(&val); err != nil {
+		return err
+	}
+	*f = NewNullFloat64(val)
+	return nil
+}
+
+// UnmarshalYAML for NullBool
+func (bb *NullBool) UnmarshalYAML(unmarshal func(interface{}) error) error {
+	var val bool
+	if err := unmarshal(&val); err != nil {
+		return err
+	}
+	*bb = NewNullBool(val)
+	return nil
+}
+
+// UnmarshalYAML for NullFloat64
+func (s *NullString) UnmarshalYAML(unmarshal func(interface{}) error) error {
+	var val string
+	if err := unmarshal(&val); err != nil {
+		return err
+	}
+	*s = NewNullString(val)
+	return nil
+}