+++ types/services/yaml.go
@@ -1,0 +1,52 @@
+package services
+
+import (
+	"github.com/pkg/errors"
+	"github.com/statping/statping/utils"
+	"gopkg.in/yaml.v2"
+)
+
+type yamlFile struct {
+	Services []*Service `yaml:"services,flow"`
+}
+
+// LoadServicesYaml will attempt to load the 'services.yml' file for Service Auto Creation on startup.
+func LoadServicesYaml() (*yamlFile, error) {
+	f, err := utils.OpenFile(utils.Directory + "/services.yml")
+	if err != nil {
+		return nil, err
+	}
+
+	var svrs *yamlFile
+	if err := yaml.Unmarshal([]byte(f), &svrs); err != nil {
+		log.Errorln("Unable to parse the services.yml file", err)
+		return nil, err
+	}
+
+	log.Infof("Found %d services inside services.yml file", len(svrs.Services))
+
+	for _, svr := range svrs.Services {
+		log.Infof("Service %s %d, hash: %s", svr.Name, svr.Id, svr.Hash())
+		if findServiceByHash(svr.Hash()) == nil {
+			if err := svr.Create(); err != nil {
+				return nil, errors.Wrapf(err, "could not create service %s", svr.Name)
+			}
+			log.Infof("Automatically created service '%s' checking %s", svr.Name, svr.Domain)
+
+			go ServiceCheckQueue(svr, true)
+		}
+	}
+
+	return svrs, nil
+}
+
+// findServiceByHas will return a service that matches the SHA256 hash of a service
+// Service hash example: sha256(name:EXAMPLEdomain:HTTP://DOMAIN.COMport:8080type:HTTPmethod:GET)
+func findServiceByHash(hash string) *Service {
+	for _, service := range All() {
+		if service.Hash() == hash {
+			return service
+		}
+	}
+	return nil
+}